Remove, and filters
Added a function that removes the artist by ID(the only part of the object that can't be the same).

Then added two functions to get artists by using genre and country. This will be used in the filtering process later on in the code.

function removeArtistById(artists, id){

    for(let i = 0; i < artists.length; i++){
        let artist = artists[i]

        if(artist.id == id){
            artist.splice(i,1);

        }

    }
}

//Two functions to get artists by using genre and country. This will be used in the
// filtering process later on in the code.

function getArtistByGenre(artists, genre){

    let artistByGenre = [];

    for(let artist of artists){
        if(artist.genre.toLowerCase() == genre.toLowerCase()){
        artistByGenre.push(artist);
        }
    }

    return artistByGenre;

}

function getArtistByCountry(artists, country){

    let artistByCountry = [];

    for(let artist of artists){
        if(artist.country.toLowerCase() == country.toLowerCase()){
        artistByCountry.push(artist);
        }
    }

    return artistByCountry;
}